fix(comments): isolate stripHtmlTags from DOMPurify/jsdom load
